Created the initial UI design
added a logo that is displayed at the top of the screen. Created a FlatList that will display the messages being sent between clients. Created a InputText that allows the user to type their message. This information is saved in a component state. Finally created a button that would send a message, not yet implemented